+++ trunk/BizHawk.Emulation/Computers/Commodore64/MOS/Chip2114.cs
@@ -31,12 +31,12 @@
 			ram[addr & 0x3FF] = (byte)(val & 0xF);
 		}
 
-		public byte Read(ushort addr)
+		public byte Read(int addr)
 		{
 			return (byte)(ram[addr & 0x3FF]);
 		}
 
-		public byte Read(ushort addr, byte bus)
+		public byte Read(int addr, byte bus)
 		{
 			return (byte)(ram[addr & 0x3FF] | (bus & 0xF0));
 		}
@@ -47,7 +47,7 @@
 			ser.Sync("ram", ref buffer);
 		}
 
-		public void Write(ushort addr, byte val)
+		public void Write(int addr, byte val)
 		{
 			ram[addr & 0x3FF] = (byte)(val & 0xF);
 		}
